+++ src/main/java/com/fangxiong/pojo/OperateLog.java
@@ -10,8 +10,8 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class OperateLog {
-    private Integer id; //ID
-    private Integer operateEmpId; //操作人ID
+    private Long id; //ID
+    private Long operateEmpId; //操作人ID
     private LocalDateTime operateTime; //操作时间
     private String className; //操作类名
     private String methodName; //操作方法名
